requirements.txt
# with Python 3.8.0
numpy==1.22.2
pyyaml==6.0
configargparse==1.5.3
tensorboard==2.8.0
fastai==2.5.3
sklearn==0.0
geojson==2.5.0
shapely==1.8.1.post1
scikit-plot==0.3.7
matplotlib==3.5.1
plotnine==0.8.0
seaborn==0.11.2
Pillow==9.0.1
opencv-python==4.5.5.62
prettytable==3.1.1
segmentation-models-pytorch==0.2.1
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.10.0+cu113
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.11.1+cu113
einops==0.4.1